Initial commit
Setup basic IPC using named pipes between the window layer (written
in C) and the application layer (written in Python).

Create the window and setup the event loop.

Use the json-c library to parse the component tree sent from the
application layer and draw it on the window using X.