Setup named pipes for IPC
IPC takes place between the app layer (written in Python) and the window layer (written in C).
The application layer notifies the window layer with latest updates in
the component tree, in MVC words, the component tree in the application layer acts as the Model, while the window in the
window layer acts as the View. The window layer notifies the application
layer about events, like the Expose event (which means that the window
changed it's size etc.) or a key press event etc. Again, in MVC words,
the window in the  window layer acts as the View, while the
application running in the application layer acts as the controller,
since it receives these events and updatea/manages application state
accordingly.